+++ changelog.md
@@ -2,14 +2,16 @@
 
 * Can't mine copper with wooden tools
 * copper nugget missing without caverns and chasms
-* slowdown aloe
-* slowdown fishing by nchanging timer
+* prevent planting mulberry
 
 ## Upcoming
 ### Additions
 * Grass no longer spreads in Autumn and Winter
 
 ### Changes
+* Enhanced fishing slowdown with seasons
+  * Now just adds time to lure the fish instead of having a chance each tick to not decrease
+  * By default, Summer adds 10s where Winter 30s
 * Lowered base breed fail chance
 * Slowed down aloe vera growth speed